cloud-course DB fixes and liquibase integration

+++ src/main/java/com/fypgrading/notificationservice/controller/NotificationController.java
@@ -25,7 +25,7 @@ public class NotificationController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Void> readNotification(@PathVariable Integer id) {
+    public ResponseEntity<Void> readNotification(@PathVariable Long id) {
         notificationService.readNotification(id);
         return ResponseEntity.ok().build();
     }
